+++ packages/block-editor/src/utils/get-paste-event-data.js
@@ -25,9 +25,9 @@ export function getPasteEventData( { clipboardData } ) {
 		}
 	}
 
-	const files = [
-		...getFilesFromDataTransfer( clipboardData ),
-	].filter( ( { type } ) => /^image\/(?:jpe?g|png|gif)$/.test( type ) );
+	const files = getFilesFromDataTransfer(
+		clipboardData
+	).filter( ( { type } ) => /^image\/(?:jpe?g|png|gif)$/.test( type ) );
 
 	// Only process files if no HTML is present.
 	// A pasted file may have the URL as plain text.