+++ packages/edit-site/src/components/editor/global-styles-provider.js
@@ -13,14 +13,22 @@ import {
 	useEffect,
 	useMemo,
 } from '@wordpress/element';
+import {
+	__EXPERIMENTAL_STYLE_PROPERTY as STYLE_PROPERTY,
+	store as blocksStore,
+} from '@wordpress/blocks';
 import { useEntityProp } from '@wordpress/core-data';
-import { __EXPERIMENTAL_STYLE_PROPERTY as STYLE_PROPERTY } from '@wordpress/blocks';
 import { useSelect, useDispatch } from '@wordpress/data';
 
 /**
  * Internal dependencies
  */
-import { default as getGlobalStyles } from './global-styles-renderer';
+import {
+	GLOBAL_CONTEXT,
+	getValueFromVariable,
+	getPresetVariable,
+} from './utils';
+import getGlobalStyles from './global-styles-renderer';
 
 const EMPTY_CONTENT = '{}';
 
@@ -58,12 +66,68 @@ export const useGlobalStylesReset = () => {
 	];
 };
 
-export default function GlobalStylesProvider( {
-	children,
-	baseStyles,
-	contexts,
-} ) {
+const extractSupportKeys = ( supports ) => {
+	const supportKeys = [];
+	Object.keys( STYLE_PROPERTY ).forEach( ( name ) => {
+		if ( get( supports, STYLE_PROPERTY[ name ].support, false ) ) {
+			supportKeys.push( name );
+		}
+	} );
+	return supportKeys;
+};
+
+const getContexts = ( blockTypes ) => {
+	const result = {
+		...GLOBAL_CONTEXT,
+	};
+
+	// Add contexts from block metadata.
+	blockTypes.forEach( ( blockType ) => {
+		const blockName = blockType.name;
+		const blockSelector = blockType?.supports?.__experimentalSelector;
+		const supports = extractSupportKeys( blockType?.supports );
+		const hasSupport = supports.length > 0;
+
+		if ( hasSupport && typeof blockSelector === 'string' ) {
+			result[ blockName ] = {
+				selector: blockSelector,
+				supports,
+				blockName,
+			};
+		} else if ( hasSupport && typeof blockSelector === 'object' ) {
+			Object.keys( blockSelector ).forEach( ( key ) => {
+				result[ key ] = {
+					selector: blockSelector[ key ].selector,
+					supports,
+					blockName,
+					title: blockSelector[ key ].title,
+					attributes: blockSelector[ key ].attributes,
+				};
+			} );
+		} else if ( hasSupport ) {
+			const suffix = blockName.replace( 'core/', '' ).replace( '/', '-' );
+			result[ blockName ] = {
+				selector: '.wp-block-' + suffix,
+				supports,
+				blockName,
+			};
+		}
+	} );
+
+	return result;
+};
+
+export default function GlobalStylesProvider( { children, baseStyles } ) {
 	const [ content, setContent ] = useGlobalStylesEntityContent();
+	const { blockTypes, settings } = useSelect( ( select ) => {
+		return {
+			blockTypes: select( blocksStore ).getBlockTypes(),
+			settings: select( 'core/edit-site' ).getSettings(),
+		};
+	} );
+	const { updateSettings } = useDispatch( 'core/edit-site' );
+
+	const contexts = useMemo( () => getContexts( blockTypes ), [ blockTypes ] );
 
 	const { userStyles, mergedStyles } = useMemo( () => {
 		const newUserStyles = content ? JSON.parse( content ) : {};
@@ -98,10 +162,11 @@ export default function GlobalStylesProvider( {
 			getStyleProperty: ( context, propertyName, origin = 'merged' ) => {
 				const styles = 'user' === origin ? userStyles : mergedStyles;
 
-				return get(
+				const value = get(
 					styles?.[ context ]?.styles,
-					STYLE_PROPERTY[ propertyName ]
+					STYLE_PROPERTY[ propertyName ].value
 				);
+				return getValueFromVariable( mergedStyles, context, value );
 			},
 			setStyleProperty: ( context, propertyName, newValue ) => {
 				const newContent = { ...userStyles };
@@ -110,50 +175,56 @@ export default function GlobalStylesProvider( {
 					contextStyles = {};
 					set( newContent, [ context, 'styles' ], contextStyles );
 				}
-				set( contextStyles, STYLE_PROPERTY[ propertyName ], newValue );
+				set(
+					contextStyles,
+					STYLE_PROPERTY[ propertyName ].value,
+					getPresetVariable(
+						mergedStyles,
+						context,
+						propertyName,
+						newValue
+					)
+				);
 				setContent( JSON.stringify( newContent ) );
 			},
 		} ),
-		[ contexts, content ]
+		[ content, mergedStyles ]
 	);
 
 	useEffect( () => {
-		if (
-			typeof contexts !== 'object' ||
-			typeof baseStyles !== 'object' ||
-			typeof userStyles !== 'object'
-		) {
-			return;
-		}
-
-		const embeddedStylesheetId = 'global-styles-inline-css';
-		let styleNode = document.getElementById( embeddedStylesheetId );
-
-		if ( ! styleNode ) {
-			styleNode = document.createElement( 'style' );
-			styleNode.id = embeddedStylesheetId;
-			document
-				.getElementsByTagName( 'head' )[ 0 ]
-				.appendChild( styleNode );
-		}
-
-		styleNode.innerText = getGlobalStyles( contexts, mergedStyles );
-	}, [ contexts, baseStyles, content ] );
-
-	const settings = useSelect( ( select ) =>
-		select( 'core/edit-site' ).getSettings()
-	);
-	const { updateSettings } = useDispatch( 'core/edit-site' );
-
-	useEffect( () => {
+		const newStyles = settings.styles.filter(
+			( style ) => ! style.isGlobalStyles
+		);
 		updateSettings( {
 			...settings,
+			styles: [
+				...newStyles,
+				{
+					css: getGlobalStyles(
+						contexts,
+						mergedStyles,
+						STYLE_PROPERTY,
+						'cssVariables'
+					),
+					isGlobalStyles: true,
+					__experimentalNoWrapper: true,
+				},
+				{
+					css: getGlobalStyles(
+						contexts,
+						mergedStyles,
+						STYLE_PROPERTY,
+						'blockStyles'
+					),
+					isGlobalStyles: true,
+				},
+			],
 			__experimentalFeatures: mapValues(
 				mergedStyles,
 				( value ) => value?.settings || {}
 			),
 		} );
-	}, [ mergedStyles ] );
+	}, [ contexts, mergedStyles ] );
 
 	return (
 		<GlobalStylesContext.Provider value={ nextValue }>